Move more generic functions to detect-utils
Move functions that can be used by more classes and functions
to the detect-utils module as utils functions.
